+++ services/dynamic-sidecar/src/simcore_service_dynamic_sidecar/models/volumes.py
@@ -1,40 +1,8 @@
 from datetime import datetime
-from enum import auto
 
 import arrow
-from models_library.utils.enums import StrAutoEnum
+from models_library.volumes import VolumeStatus
 from pydantic import BaseModel, Field
-
-
-class VolumeCategory(StrAutoEnum):
-    """
-    These uniquely identify volumes which are mounted by
-    the dynamic-sidecar and user services.
-    This is primarily used to keep track of the status of
-    each individual volume on the volumes.
-    The status is ingested by the agent and processed
-    when the volume is removed.
-    """
-
-    # contains data relative to output ports
-    OUTPUTS = auto()
-
-    # contains data relative to input ports
-    INPUTS = auto()
-
-    # contains files which represent the state of the service
-    # usually the user's workspace
-    STATES = auto()
-
-    # contains dynamic-sidecar data required to maintain state
-    # between restarts
-    SHARED_STORE = auto()
-
-
-class VolumeStatus(StrAutoEnum):
-    CONTENT_NEEDS_TO_BE_SAVED = auto()
-    CONTENT_WAS_SAVED = auto()
-    CONTENT_NO_SAVE_REQUIRED = auto()
 
 
 class VolumeState(BaseModel):
