+++ services/dynamic-sidecar/tests/unit/test_api_volumes.py
@@ -5,13 +5,10 @@ from pathlib import Path
 import pytest
 from async_asgi_testclient import TestClient
 from fastapi import status
+from models_library.volumes import VolumeCategory, VolumeStatus
 from simcore_service_dynamic_sidecar._meta import API_VTAG
 from simcore_service_dynamic_sidecar.models.shared_store import SharedStore
-from simcore_service_dynamic_sidecar.models.volumes import (
-    VolumeCategory,
-    VolumeState,
-    VolumeStatus,
-)
+from simcore_service_dynamic_sidecar.models.volumes import VolumeState
 
 
 @pytest.mark.parametrize(
