+++ services/dynamic-sidecar/tests/unit/test_models_volumes.py
@@ -1,8 +1,9 @@
 # pylint: disable=redefined-outer-name
 
 import pytest
+from models_library.volumes import VolumeStatus
 from pytest import FixtureRequest
-from simcore_service_dynamic_sidecar.models.volumes import VolumeState, VolumeStatus
+from simcore_service_dynamic_sidecar.models.volumes import VolumeState
 
 
 @pytest.fixture(params=VolumeStatus)
